feat: ボスバー、スコアボードがもやんのAPI変更に伴い動かなくなっていたので修正。CarryTntのコアロジックのバグ修正。ゲームエントリーロジックのバグ修正、イベント回りを再度整え、configが正常に読み込まれるように細部を修正、ワールド生成ロジックをフォルダを丸ごとコピーに変更し軽量化、ジェネレーターとレベルシステムのバグ修正、Koinモジュールで誤解しており、正常に動作しなかった部分の修正、paper-plugin.ymlがplugin.ymlと比べて依存の書き方が変わっていたので修正、ユーティリティサービスの修正と細かな便利メソッドの追加。その他いろいろ

+++ src/main/kotlin/com/github/ryuzu/ryuzutechnicalmagiccore/core/util/wrapper/bossbar/IBossBarService.kt
@@ -13,6 +13,6 @@ interface IBossBarService : KoinComponent {
     fun removePlayers(players: Set<IPlayer>)
     fun removePlayers(vararg players: IPlayer) = removePlayers(players.toSet())
     fun destroy()
-    fun progress(progress: Double)
+    fun progress(progress: Float)
     fun color(color: String)
 }
